+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 HTTP client extension Change Log
 
 - Bug #44: Fixed exception name collision at `Response` and `Transport` (cebe)
 - Bug #45: Fixed `XmlFormatter` unable to handle array with numeric keys (klimov-paul)
+- Enh #47: Added `Message::addData()` allowing addition of the content data to already existing one (klimov-paul)
 - Enh: Added `XmlFormatter::useTraversableAsArray` allowing processing `\Traversable` as array (klimov-paul)